Fix a bug with double precision.

+++ library/ruler/src/main/java/io/mehow/ruler/ImperialLengthUnit.kt
@@ -17,8 +17,8 @@ enum class ImperialLengthUnit(
     }
   };
 
-  override fun toDistance(value: Double): Distance {
-    return Distance.create(value.toBigDecimal() * meterRatio)
+  override fun toDistance(value: BigDecimal): Distance {
+    return Distance.create(value * meterRatio)
   }
 
   override fun toMeasuredLength(meters: BigDecimal): BigDecimal {
